added support to jakartaee for scheduled annotation plugin
