+++ apm-agent-plugins/apm-scheduled-annotation-plugin/src/main/java/co/elastic/apm/agent/scheduled/ScheduledTransactionNameInstrumentation.java
@@ -41,7 +41,7 @@ import java.util.Collection;
 import static co.elastic.apm.agent.bci.bytebuddy.CustomElementMatchers.isInAnyPackage;
 import static net.bytebuddy.matcher.ElementMatchers.declaresMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isAnnotatedWith;
-import static net.bytebuddy.matcher.ElementMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.namedOneOf;
 
 public class ScheduledTransactionNameInstrumentation extends TracerAwareInstrumentation {
 
@@ -57,7 +57,7 @@ public class ScheduledTransactionNameInstrumentation extends TracerAwareInstrume
         @Nullable
         @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
         public static Object setTransactionName(@SimpleMethodSignature String signature,
-                                                 @Advice.Origin Class<?> clazz) {
+                                                @Advice.Origin Class<?> clazz) {
             AbstractSpan<?> active = tracer.getActive();
             if (active == null) {
                 Transaction transaction = tracer.startRootTransaction(clazz.getClassLoader());
@@ -95,11 +95,8 @@ public class ScheduledTransactionNameInstrumentation extends TracerAwareInstrume
     @Override
     public ElementMatcher<? super MethodDescription> getMethodMatcher() {
         return isAnnotatedWith(
-            named("org.springframework.scheduling.annotation.Scheduled")
-                .or(named("org.springframework.scheduling.annotation.Schedules"))
-                .or(named("javax.ejb.Schedule"))
-                .or(named("javax.ejb.Schedules"))
-        );
+            namedOneOf("org.springframework.scheduling.annotation.Scheduled", "org.springframework.scheduling.annotation.Schedules",
+                "javax.ejb.Schedule", "javax.ejb.Schedules", "jakarta.ejb.Schedule", "jakarta.ejb.Schedules"));
     }
 
     @Override
