+++ apm-agent-plugins/pom.xml
@@ -65,6 +65,7 @@
         <module>apm-sparkjava-plugin</module>
         <module>apm-javalin-plugin</module>
         <module>apm-servlet-jakarta-test</module>
+        <module>apm-scheduled-annotation-plugin-jakartaee-test</module>
     </modules>
 
     <dependencies>
